separation authentication to module
