+++ modules/@ergonode/authentication/src/components/Infographic/Infographic.vue
