+++ modules/@ergonode/core/src/config/routes.js
@@ -10,8 +10,6 @@ import {
 import Privileges from './privileges';
 
 export const ROUTE_NAME = {
-    INDEX: 'index',
-    PASSWORD_RESET: 'password-reset',
     PLACEHOLDER: 'placeholder',
     SETTINGS: 'settings',
     SETTINGS_LANGUAGES: 'settings-languages',
@@ -22,22 +20,6 @@ export const ROUTE_NAME = {
 };
 
 export default [
-    {
-        name: ROUTE_NAME.INDEX,
-        path: '/',
-        component: Pages.Login,
-        meta: {
-            isMenu: false,
-        },
-    },
-    {
-        name: ROUTE_NAME.PASSWORD_RESET,
-        path: '/password-reset',
-        component: Pages.PasswordReset,
-        meta: {
-            isMenu: false,
-        },
-    },
     {
         name: ROUTE_NAME.PLACEHOLDER,
         path: '/placeholder/:placeholder',
